chan_sip: Revert removal.
This is a partial reversion of Asterisk commit
4095a382da630f8af2aac2d223fcaa70c03b81ba
which is the commit that removed chan_sip
from Asterisk.

The reversion mainly consists of readding
entire files that were removed. A diff
is included for a Makefile change.

Many other files were altered in the
aforementioned commit that are not included
here. Many of these are only tangentially
relevant (e.g. references to chan_sip in
other files), and these are not included
in the reversion to minimize future conflicts
and keep chan_sip as self-contained as possible.